+++ ErsatzTV.Core/Metadata/LanguageCodeAndName.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Core.Metadata;
+
+public record LanguageCodeAndName(string Code, string Name);