+++ CHANGELOG.md
@@ -44,6 +44,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix adding pad filler to content that is less than one minute in duration
 - Generate unique identifier for virtual HDHomeRun tuner by @raknam
   - This allows a single Plex server to connect to multiple ETV instances
+- Include *all* language codes from media library in preferred audio and subtitle language options
+  - Language codes where an English name cannot be found will be at the bottom of the list
 
 ### Changed
 - Remove some unnecessary API calls related to media server scanning and paging